requirements.txt
streamlit
transformers
datasets
torch
accelerate
streamlit-audiorecorder